requirements.txt
absl-py==2.0.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.2
altgraph==0.17.4
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.35.4
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
blis==0.7.11
bypy==1.8.5
cachetools==5.3.2
captum==0.7.0
catalogue==2.0.10
clarabel==0.9.0
click==8.1.7
clip==0.2.0
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
colorlover==0.3.0
confection==0.1.4
contourpy==1.1.1
cpm-kernels==1.0.11
crc32c==2.3.post0
crcmod==1.7
cvxpy==1.5.2
cycler==0.12.1
cymem==2.0.8
data==0.4
datasets==2.15.0
decord==0.6.0
defusedxml==0.7.1
diffusers==0.19.3
dill==0.3.7
docopt==0.6.2
easydict==1.13
ecos==2.0.14
einops==0.7.0
exceptiongroup==1.1.3
fairscale==0.4.13
fastapi==0.104.1
fastjsonschema==2.19.1
fasttext==0.9.2
ffmpy==0.3.1
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
funcsigs==1.0.2
fvcore==0.1.5.post20221221
gast==0.4.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1.1
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.30.2
humanfriendly==10.0
icecream==2.1.0
imageio==2.34.0
imageio-ffmpeg==0.5.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
ipython==8.12.3
jieba==0.42.1
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyterlab_pygments==0.3.0
keras==2.13.1
kiwisolver==1.4.5
labelImg==1.8.6
langcodes==3.3.0
latex2mathml==3.76.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.11.7
llvmlite==0.41.1
loralib==0.1.2
lpips==0.1.4
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
matplotlib==3.7.3
mdtex2html==1.2.0
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
modelcards==0.1.6
modelscope==1.13.1
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nerfacc==0.3.3
netron==7.4.5
ninja==1.11.1.1
nltk==3.8.1
numba==0.58.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.15.0
onnxruntime==1.19.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.9.10
osqp==0.6.7.post1
oss2==2.18.4
paddlepaddle==2.6.1
pandas==2.0.3
pandocfilters==1.5.1
pipreqs==0.5.0
piq==0.8.0
plotly==5.18.0
pooch==1.8.2
portalocker==2.10.1
POT==0.9.5
preshed==3.0.9
prettytable==3.10.0
properscoring==0.1
protobuf==4.25.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pyby==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycryptodome==3.20.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyhocon==0.3.57
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMatting==1.1.12
PyMCubes==0.1.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyransac3d==0.6.0
pyre-extensions==0.0.23
python-multipart==0.0.6
pytorch-lightning==1.9.5
pytorch3d==0.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
pyzmq==19.0.2
qdldl==0.1.7.post4
referencing==0.30.2
regex==2023.10.3
rembg==2.0.59
rich==13.6.0
rouge-chinese==1.0.3
rpds-py==0.12.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.7
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.19.2
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sse-starlette==1.6.5
starlette==0.27.0
streamlit==1.28.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tfrecord==1.14.4
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.5.4
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torchdata==0.7.1
torchmetrics==1.4.1
torchsummary==1.5.1
torchtext==0.16.1
tqdm==4.66.1
transformers==4.46.3
trimesh==3.9.8
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uvicorn==0.24.0.post1
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
yarl==1.9.3
zipp==3.17.0
